Refactor to remove `constants.ts` (#978)
