+++ src/processTargets/modifiers/subToken.ts
@@ -1,0 +1,1 @@
+export const SUBWORD_MATCHER = /[A-Z]?[a-z]+|[A-Z]+(?![a-z])|[0-9]+/g;
