+++ src/scripts/hatAdjustments/add.ts
@@ -2,7 +2,7 @@
  * This script can be used to add hat tweaks to the currently shipping ones
  */
 import { sum } from "lodash";
-import { HAT_SHAPES } from "../../core/constants";
+import { HAT_SHAPES } from "../../core/hatStyles";
 import {
   defaultShapeAdjustments,
   HatAdjustments,
